ml_server/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
pillow
pydantic
